Introduce the ability for the shell to run a command when a c# event is raised
The user can use the 'event' command to execute a command when a C#
event is raised. The event must have the 'UShell.Event' attribute to
be accessible through the 'event' command.
The 'API Compatibility Level' has been set to .NET 4.x (was .NET
Standard 2.0) in the project settings to allow the above
functionnality. The 'setVariableValue' and 'getVariableValue'
method from 'Shell.cs' are now public instead of private.